+++ forc-plugins/forc-fmt-v2/Cargo.toml
@@ -1,10 +1,11 @@
 [package]
 name = "forc-fmt-v2"
-version = "0.14.4"
+version = "0.0.0"
 authors = ["Fuel Labs <contact@fuel.sh>"]
 edition = "2021"
 homepage = "https://fuel.network/"
 license = "Apache-2.0"
+publish = false
 repository = "https://github.com/FuelLabs/sway"
 description = "A `forc` plugin for running the Sway code formatter."
